Move LoginResponse to response/ folder to be consistent with Lucy's structure
