+++ capstone/src/main/java/com/google/sps/servlets/LoginServlet.java
@@ -14,7 +14,7 @@
 
 package com.google.sps.servlets;
 
-import com.google.sps.Objects.LoginResponse;
+import com.google.sps.Objects.response.LoginResponse;
 import com.google.gson.Gson;
 import com.google.appengine.api.users.UserService;
 import com.google.appengine.api.users.UserServiceFactory;
